requirements.txt
python-telegram-bot==20.3
Flask==2.3.2
python-dotenv==1.0.0
requests==2.31.0
flask[async]
